# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/summary": {
      "get": {
        "operationId": "AdminController_summary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/kie-economics": {
      "get": {
        "operationId": "AdminController_getKieEconomics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_patchKieEconomics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putKieEconomics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/openai-vless": {
      "get": {
        "operationId": "AdminController_getVless",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putVless",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/apimart": {
      "get": {
        "operationId": "AdminController_getApimart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putApimart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/events": {
      "get": {
        "operationId": "AdminController_events",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/support/threads": {
      "get": {
        "operationId": "AdminController_supportThreads",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/support/{userId}/messages": {
      "get": {
        "operationId": "AdminController_supportMessages",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_supportMessagesPost",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/support/{userId}/read": {
      "post": {
        "operationId": "AdminController_supportRead",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/support/{userId}/close": {
      "post": {
        "operationId": "AdminController_supportClose",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/promo-codes": {
      "get": {
        "operationId": "AdminController_promoCodes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_createPromo",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/grant": {
      "post": {
        "operationId": "AdminController_grantUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/search": {
      "get": {
        "operationId": "AdminController_searchUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{userId}": {
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/kie-photo-prices": {
      "get": {
        "operationId": "AdminController_photoPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putPhotoPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/kie-video-prices": {
      "get": {
        "operationId": "AdminController_videoPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_putVideoPrices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/models": {
      "get": {
        "operationId": "AdminController_models",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/models/{id}": {
      "put": {
        "operationId": "AdminController_putModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/models/{id}/active": {
      "patch": {
        "operationId": "AdminController_patchModelActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/publish-bans": {
      "get": {
        "operationId": "AdminController_publishBans",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/feed/remove": {
      "post": {
        "operationId": "AdminController_feedRemove",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/publish-bans/restore": {
      "post": {
        "operationId": "AdminController_bansRestore",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/publish-bans/dismiss": {
      "post": {
        "operationId": "AdminController_bansDismiss",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/banners/{slot}": {
      "post": {
        "operationId": "AdminController_upsertBanner",
        "parameters": [
          {
            "name": "slot",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteBanner",
        "parameters": [
          {
            "name": "slot",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/applications": {
      "get": {
        "operationId": "AdminController_partnerApplications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/applications/{id}/approve": {
      "post": {
        "operationId": "AdminController_approveApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/applications/{id}/reject": {
      "post": {
        "operationId": "AdminController_rejectApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/list": {
      "get": {
        "operationId": "AdminController_partnerList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/payouts": {
      "get": {
        "operationId": "AdminController_partnerPayouts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/{id}": {
      "get": {
        "operationId": "AdminController_partnerDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_patchPartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/{id}/sub-partners": {
      "post": {
        "operationId": "AdminController_addSubPartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/{id}/sub-partners/{childId}": {
      "delete": {
        "operationId": "AdminController_removeSubPartner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "childId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/partner/payouts/{id}/process": {
      "post": {
        "operationId": "AdminController_processPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Neiro API",
    "description": "OpenAPI 3 спецификация, сгенерированная из NestJS-контроллеров (@nestjs/swagger).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Authorization: Bearer <token> (как в Express-сервере)"
      }
    },
...
  }
}